changed font, fixed knight bug

+++ src/util/linalg.c
@@ -2,7 +2,7 @@
 #include <math.h>
 #include <stdio.h>
 
-vec2 vec2_create(f32 x, f32 y)
+vec2 vec2_create(f64 x, f64 y)
 {
     return (vec2) { .x = x, .y = y };
 }
@@ -19,24 +19,24 @@ vec2 vec2_sub(vec2 v1, vec2 v2)
 
 vec2 vec2_normalize(vec2 vec)
 {
-    f32 mag = vec2_mag(vec);
+    f64 mag = vec2_mag(vec);
     if (mag == 0)
         return vec;
     return (vec2) { .x = vec.x / mag, .y = vec.y / mag };
 
 }
 
-vec2 vec2_scale(vec2 vec, f32 scale)
+vec2 vec2_scale(vec2 vec, f64 scale)
 {
     return (vec2) { .x = vec.x * scale, .y = vec.y * scale };
 }
 
-vec2 vec2_direction(f32 rad)
+vec2 vec2_direction(f64 rad)
 {
     return (vec2) { .x = cos(rad), .y = sin(rad) };
 }
 
-vec2 vec2_rotate(vec2 vec, f32 rad)
+vec2 vec2_rotate(vec2 vec, f64 rad)
 {
     return (vec2) { 
         .x = vec.x * cos(rad) - vec.y * sin(rad),
@@ -52,25 +52,25 @@ vec2 vec2_rotate180(vec2 vec)
     };
 }
 
-f32 vec2_radians(vec2 vec)
+f64 vec2_radians(vec2 vec)
 {
     // atan2f does this
     return atan(vec.y / vec.x) + ((vec.x < 0) ? PI : 0);
 }
 
-f32 vec2_mag(vec2 vec)
+f64 vec2_mag(vec2 vec)
 {
     return sqrt(vec.x * vec.x + vec.y * vec.y);
 }
 
-f32 vec2_distance(vec2 v1, vec2 v2)
+f64 vec2_distance(vec2 v1, vec2 v2)
 {
-    f32 dx = v2.x - v1.x;
-    f32 dy = v2.y - v1.y;
+    f64 dx = v2.x - v1.x;
+    f64 dy = v2.y - v1.y;
     return sqrt(dx * dx + dy * dy);
 }
 
-f32 vec2_dot(vec2 v1, vec2 v2)
+f64 vec2_dot(vec2 v1, vec2 v2)
 {
     return v1.x * v2.x + v1.y * v2.y;
 }
@@ -85,7 +85,7 @@ void vec2_print(vec2 vec)
     printf("(%.5f,%.5f)\n", vec.x, vec.y);
 }
 
-vec3 vec3_create(f32 x, f32 y, f32 z)
+vec3 vec3_create(f64 x, f64 y, f64 z)
 {
     return (vec3) { x, y, z };
 }
@@ -102,13 +102,13 @@ vec3 vec3_sub(vec3 v1, vec3 v2)
 
 vec3 vec3_normalize(vec3 vec)
 {
-    f32 mag = vec3_mag(vec);
+    f64 mag = vec3_mag(vec);
     if (mag == 0)
         return vec;
     return (vec3) { vec.x / mag, vec.y / mag, vec.z / mag };
 }
 
-vec3 vec3_scale(vec3 vec, f32 scale)
+vec3 vec3_scale(vec3 vec, f64 scale)
 {
     return (vec3) { vec.x * scale, vec.y * scale, vec.z * scale };
 }
@@ -122,12 +122,12 @@ vec3 vec3_cross(vec3 v1, vec3 v2)
     };
 }
 
-f32 vec3_dot(vec3 v1, vec3 v2)
+f64 vec3_dot(vec3 v1, vec3 v2)
 {
     return v1.x * v2.x + v1.y * v2.y + v1.z * v2.z;
 }
 
-f32 vec3_mag(vec3 vec)
+f64 vec3_mag(vec3 vec)
 {
     return sqrt(vec.x * vec.x + vec.y * vec.y + vec.z * vec.z);
 }
@@ -162,26 +162,26 @@ OBB obb_create(vec3 origin, vec3 offset, vec3 orientation)
     return (OBB) { origin, offset, orientation };
 }
 
-f32 intersect_ray_triangle(Ray ray, Triangle triangle)
+f64 intersect_ray_triangle(Ray ray, Triangle triangle)
 {
     // Möller–Trumbore
     vec3 edge1 = vec3_sub(triangle.b, triangle.a);
     vec3 edge2 = vec3_sub(triangle.c, triangle.a);
     vec3 ray_cross_e2 = vec3_cross(ray.direction, edge2);
-    f32 det = vec3_dot(edge1, ray_cross_e2);
+    f64 det = vec3_dot(edge1, ray_cross_e2);
 
     if (det > -EPSILON && det < EPSILON)
         return -INF;
 
-    f32 inv_det = 1 / det;
+    f64 inv_det = 1 / det;
     vec3 s = vec3_sub(ray.origin, triangle.a);
-    f32 u = inv_det * vec3_dot(s, ray_cross_e2);
+    f64 u = inv_det * vec3_dot(s, ray_cross_e2);
 
     if ((u < 0 && fabs(u) > EPSILON) || (u > 1 && fabs(u - 1) > EPSILON))
         return -INF;
 
     vec3 s_cross_e1 = vec3_cross(s, edge1);
-    f32 v = inv_det * vec3_dot(ray.direction, s_cross_e1);
+    f64 v = inv_det * vec3_dot(ray.direction, s_cross_e1);
 
     if ((v < 0 && fabs(v) > EPSILON) || (u + v > 1 && fabs(u + v - 1) > EPSILON))
         return -INF;
@@ -197,25 +197,25 @@ bool intersect_segment_triangle(Segment segment, Triangle triangle)
     vec3 edge1 = vec3_sub(triangle.b, triangle.a);
     vec3 edge2 = vec3_sub(triangle.c, triangle.a);
     vec3 ray_cross_e2 = vec3_cross(ray_direction, edge2);
-    f32 det = vec3_dot(edge1, ray_cross_e2);
+    f64 det = vec3_dot(edge1, ray_cross_e2);
 
     if (det > -EPSILON && det < EPSILON)
         return false;
 
-    f32 inv_det = 1 / det;
+    f64 inv_det = 1 / det;
     vec3 s = vec3_sub(ray_origin, triangle.a);
-    f32 u = inv_det * vec3_dot(s, ray_cross_e2);
+    f64 u = inv_det * vec3_dot(s, ray_cross_e2);
 
     if ((u < 0 && fabs(u) > EPSILON) || (u > 1 && fabs(u - 1) > EPSILON))
         return false;
 
     vec3 s_cross_e1 = vec3_cross(s, edge1);
-    f32 v = inv_det * vec3_dot(ray_direction, s_cross_e1);
+    f64 v = inv_det * vec3_dot(ray_direction, s_cross_e1);
 
     if ((v < 0 && fabs(v) > EPSILON) || (u + v > 1 && fabs(u + v - 1) > EPSILON))
         return false;
     
-    f32 t = inv_det * vec3_dot(edge2, s_cross_e1);
+    f64 t = inv_det * vec3_dot(edge2, s_cross_e1);
     return t > -EPSILON && t < 1 + EPSILON;
 }
 
@@ -233,9 +233,9 @@ bool intersect_aabb_hexahedron(AABB aabb, Hexahedron hexahedron)
 {
     vec3 points[8];
     vec3 normal;
-    f32 distance;
+    f64 distance;
     i32 i, j;
-    f32 x, y, z, dx, dy, dz;
+    f64 x, y, z, dx, dy, dz;
     x = aabb.origin.x; y = aabb.origin.y; z = aabb.origin.z;
     dx = aabb.offset.x; dy = aabb.offset.y; dz = aabb.offset.z;
     points[0] = vec3_create(x     , y     , z     );
